Refactored the PS1 memcard stuff into its own class
